Refactor: Move DetailsPane to panes.details
OPEN - task 54: Create Category CRUD packages 
http://github.com/DevOpsDistilled/OpERP/issues/issue/54
